services/risk-assessment-service/ml-training/requirements.txt
# PyTorch and ML dependencies
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyarrow>=10.0.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter and development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Model evaluation
tensorboard>=2.13.0
wandb>=0.15.0

# Data validation
great-expectations>=0.17.0